fix merge

+++ include/RAG.h
@@ -28,7 +28,7 @@ typedef struct cellule_t Cellule;
 
 extern RAG* create_RAG(image im, int n, int m);
 
-extern double RAG_give_closest_region(RAG* rag, int*, int*);
+extern long double RAG_give_closest_region(RAG* rag, int*, int*);
 
 void RAG_merge_region(RAG * rag, int i, int j);
